% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lin_ah.R
\name{lin_ah}
\alias{lin_ah}
\title{Additive hazards regression model}
\usage{
lin_ah(
  time,
  event,
  covariates,
  weights = NULL,
  offset = rep(0, length(time)),
  variance = TRUE
)
}
\arguments{
\item{time}{vector of time-to-event}

\item{event}{vector of event indicator; 0 = censored}

\item{covariates}{design matrix of covariates; intercept should not be included}

\item{offset}{n-vector of offset}

\item{variance}{whether to return the variance and components for the calculation}
}
\description{
Fit semiparametric additive hazards model in Lin & Ying 1997 with time-independent
covariates.
This is a utility function.
}
\examples{
x <- rep(1:3, each = 10)
t_uncensored <- rexp(30, rate = 1 / exp(0.5 -0.1 * x))
t_censored <- pmin(5, t_uncensored)
event <- as.numeric(t_uncensored <= 5)

ah_result <- lin_ah(time = t_censored, event = event, covariates = x)
ah_result$summary
}
