% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_fit.R
\name{linear_fit}
\alias{linear_fit}
\title{Linear model}
\usage{
linear_fit(y, x, offset = rep(0, length(y)), variance = TRUE)
}
\arguments{
\item{y}{n-vector of outcome}

\item{x}{design matrix}

\item{offset}{n-vector of offset}

\item{variance}{whether to return the variance and components for the calculation}
}
\description{
Fit linear model E[Y | X] = offset + beta * x
}
