% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logit_reg.R
\name{logit_reg}
\alias{logit_reg}
\title{Logistic regression model model}
\usage{
logit_reg(y, x, offset = rep(0, length(y)), variance = TRUE)
}
\arguments{
\item{y}{n-vector of binary outcome}

\item{x}{design matrix}

\item{offset}{n-vector of offset}

\item{variance}{whether to return the variance and components for the calculation}
}
\description{
Fit linear model E[Y | X] = expit(offset + beta * x)
}
\examples{
x <- rep(1:3, each = 10)
y <- rbinom(30, 1, 1 / (1 + exp(1 - 2 * x)))

logitreg_result <- logit_reg(y, cbind(1, x))
logitreg_result$summary
}
