% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglin.R
\name{negbin_fit}
\alias{negbin_fit}
\title{Negative-binomial regression model}
\usage{
negbin_fit(
  y,
  x,
  offset = rep(0, length(y)),
  variance = TRUE,
  interval = c(-100, 100)
)
}
\arguments{
\item{y}{n-vector of outcome}

\item{x}{design matrix}

\item{offset}{n-vector of offset}

\item{variance}{whether to return the variance and components for the calculation}
}
\description{
Fit negative-binomial regression model Y | X ~ NegativeBinomial(theta, exp(offset + beta * x))
}
