% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pciah.R
\name{p2sls.ah}
\alias{p2sls.ah}
\title{Proximal Two-stage-least-squares with survival outcome}
\usage{
p2sls.ah(
  Y,
  D,
  A,
  X = NULL,
  W,
  Z = NULL,
  Xw = NULL,
  Xy = NULL,
  nco_type = NULL,
  nco_args = NULL,
  variance = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{Y}{an n-vector of observed time to event outcomes}

\item{D}{an n-vector of event indicators for Y}

\item{A}{an n-vector of primary exposure}

\item{X}{an n * nX matrix of adjusted covariates , can be empty}

\item{W}{an n * nW matrix of negative control outcome (NCO)}

\item{Z}{an n * nZ matrix of negative control exposure (NCE). Alternatively,
the NCE can be included in the design matrix of the first-stage model by specifying
Xw}

\item{Xw}{a matrix with n rows or a list of length nW, each element is a design matrix of
the corresponding NCO. Default is a matrix with columns of A, Z, X, and interactions between
A, Z and A, X}

\item{Xy}{a matrix with n rows, covariates to be adjusted for in the second-stage
model besides A and predictors for W. This may include interaction between A and columns of X.
Default to be X}

\item{nco_type}{An nW-vector for the types of models used in the first-stage, including
"linear" (linear model), "loglin" (log-linear model), "poisson" (Poisson regression model),
"negbin" (negative-binomial regression model), and "ah" (Lin \& Ying's additive hazards model)}

\item{nco_args}{A list of length nW containing additional arguments for the first-stage models.
Each element in the list is a sublist and should contain a vector named "offset" (default 0).
If nco_type == "ah", the sublist needs to include another vector named "event" as the event indicator
(default 1). If nco_type == "negbin", the sublist needs to include a vector "init" as the initial values,
which can be NA.}

\item{variance}{whether to return the variance and components for the calculation}
}
\value{
A list with three elements: ESTIMATE includes the parameter estimates from the second-stage
model; SE produces their standard error, and PARAM includes all parameter estimates
}
\description{
This function computes the adjusted hazard difference using additive hazards model by
Lin \& Ying 1994, using a pair of negative control variable to control for unmeasured
confounding
}
\examples{
N <- 2000
U <- runif(N); X <- runif(N)
expit  <-  function(x) exp(x)/(1 + exp(x))
A <- rbinom(N, 1, expit(-3 + 5 * U + 1 * X))
Y <- rexp(N, 0.2 + 1.5 * U + 0.2 * X + 0.2 * A)
D <- as.numeric(Y < 5)
Y[D == 0] <- 5
Z <- rnorm(N, 2 * U + 0.5 * X)
W2 <- rexp(N, 0.1 + 1 * U + 0.1 * X)
D2 <- as.numeric(W2 < 5)
W2[D2 == 0] <- 5
W <- cbind(W1 = rnbinom(N, size = 25, mu = exp(2.5 * U + 0.2 * X)),
           W2)
p2sls_result <- p2sls.ah(Y = Y, D = D, A = A, X = X,
       W = W, Z = Z, variance = TRUE,
       nco_type = c("negbin", "ah"),
       nco_args = list(list(offset = rep(0, N)),
                       list(offset = rep(0, N),
                            event = D2)))
p2sls_result$summary_first_stage
p2sls_result$summary_second_stage

}
