% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p2sls.cprisk.nc.R
\name{p2sls.cprisk.nc}
\alias{p2sls.cprisk.nc}
\title{Counterfactual marginal CIF for proximal two-stage-least-squares with survival outcome using competing risks as negative control outcomes, but the exposure has an effect to the competing risk during the initial period after the treatment}
\usage{
p2sls.cprisk.nc(
  times,
  cause,
  A,
  X,
  Z,
  nc_time,
  bootstrap = F,
  nboot = 1000,
  conf.level = 0.95
)
}
\arguments{
\item{times}{an n-vector of observed time to event outcomes}

\item{cause}{an n-vector of indicators for causes of events; in each entry, 0 indicates primary event, 1 indicates competing risks, and -1 indicates censoring}

\item{A}{an n-vector of primary exposure}

\item{X}{an n * nX matrix of adjusted covariates , can be empty}

\item{Z}{an n * nZ matrix of negative control exposure (NCE)}

\item{nc_time}{time after which the exposure has no effect on the competing risk}
}
\value{
A data frame with two columns: t: time points where the counterfactual marginal survival functions are evaluated; 
survfunc: value of the counterfactual marginal survival function
}
\description{
This function computes the counterfactual marginal cumulative cause-specific hazard functions and cumulative incidence functions for the proximal 
two-stage-least-squares with survival outcome, using competing risks as the negative control outcomes
}
\examples{
set.seed(2050)
N <- 2000
U <- runif(N); X <- runif(N)
expit  <-  function(x) exp(x)/(1 + exp(x))
A <- rbinom(N, 1, expit(-3 + 5 * U + 1 * X))
Y1 <- rexp(N, 0.5 * U + 0.2 * X + 0.2 * A)  ## event times for primary event of interest
Y2 <- rexp(N, 0.3 * U + 0.6 * X)  ## event times for the competing risk
C <- runif(N, 3, 5)  ## censoring times
times <- pmin(Y1, Y2, C)
cause <- ifelse(Y1 <= C, ifelse(Y1 <= Y2, 0, 1), ifelse(Y2 <= C, 1, -1))
Z <- rnorm(N, 2 * U + 0.5 * X)
## Obtain the counterfactual marginal survival curves under a = 0, 1

p2sls_rslt <- p2sls.cprisk.nc(times = times, cause = cause, nc_time = 2, A = A, X = X, Z = Z, bootstrap = T, nboot = 100)
}
