% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p2sls.cprisk.R
\name{p2sls.cprisk}
\alias{p2sls.cprisk}
\title{Proximal two-stage-least-squares with survival outcome using competing risks as negative control outcomes}
\usage{
p2sls.cprisk(
  times,
  cause,
  A,
  a,
  X,
  Z,
  surv_correct = T,
  tmax = NULL,
  nt = 1000
)
}
\arguments{
\item{times}{an n-vector of observed time to event outcomes}

\item{cause}{an n-vector of indicators for causes of events; in each entry, 0 indicates primary event, 1 indicates competing risks, and -1 indicates censoring}

\item{A}{an n-vector of primary exposure}

\item{a}{a unidimensional scalar value to fix A=a for calculating the counterfactual survival curve}

\item{X}{an n * nX matrix of adjusted covariates , can be empty}

\item{Z}{an n * nZ matrix of negative control exposure (NCE)}

\item{surv_correct}{logical, whether the survival function will be corrected post hoc to be non-increasing and
fall between 0 and 1, default to be TRUE.}

\item{tmax}{The maximal time to calculate the counterfactual marginal survival functions. Default to be the 
maximal event time multiplied by 1.05;}

\item{nt}{Number of time points to evaluate the survival functions, evenly distributed between 0 and tmax.}
}
\value{
A data frame with two columns: t: time points where the counterfactual marginal survival functions are evaluated; 
survfunc: value of the counterfactual marginal survival function
}
\description{
This function computes the counterfactual marginal cumulative cause-specific hazard functions and cumulative incidence functions for the proximal 
two-stage-least-squares with survival outcome, using competing risks as the negative control outcomes
}
\examples{
set.seed(2050)
N <- 2000
U <- runif(N); X <- runif(N)
expit  <-  function(x) exp(x)/(1 + exp(x))
A <- rbinom(N, 1, expit(-3 + 5 * U + 1 * X))
Y1 <- rexp(N, 0.5 * U + 0.2 * X + 0.2 * A)  ## event times for primary event of interest
Y2 <- rexp(N, 0.3 * U + 0.6 * X)  ## event times for the competing risk
C <- runif(N, 3, 5)  ## censoring times
times <- pmin(Y1, Y2, C)
cause <- ifelse(Y1 <= C, ifelse(Y1 <= Y2, 0, 1), ifelse(Y2 <= C, 1, -1))
Z <- rnorm(N, 2 * U + 0.5 * X)
## Obtain the counterfactual marginal survival curves under a = 0, 1

p2sls_rslt_a1 <- p2sls.cprisk(times = times, cause = cause, A = A, a = 1, X = X, Z = Z)
p2sls_rslt_a0 <- p2sls.cprisk(times = times, cause = cause, A = A, a = 0, X = X, Z = Z)


## Plot the counterfactual marginal cause-specific cumulative hazard function for the primary event 
plot(cumhaz0 ~ t, data = p2sls_rslt_a1, type = "l", lwd = 2,
    xlab = "Time", ylab = "Cumulative hazard function", col = "red")
lines(cumhaz0 ~ t, data = p2sls_rslt_a0, lwd = 2) 
legend("topleft", legend = c("a = 0", "a = 1"), lwd = 2,
      col = c("black", "red"))
      
## Plot the counterfactual marginal cumulative incidence functions for the primary event 
plot(cif0 ~ t, data = p2sls_rslt_a1, type = "l", lwd = 2,
    xlab = "Time", ylab = "Cumulative incidence", col = "red")
lines(cif0 ~ t, data = p2sls_rslt_a0, lwd = 2) 
legend("topleft", legend = c("a = 0", "a = 1"), lwd = 2,
      col = c("black", "red"))     
      
}
