% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcilogit.R
\name{pci.logitreg}
\alias{pci.logitreg}
\title{Proximal causal inference with categorical outcomes using two-stage-least-square}
\usage{
pci.logitreg(
  Y,
  offset = rep(0, length(Y)),
  A,
  X = NULL,
  W,
  Z = NULL,
  Xw = NULL,
  Xw_Y0 = NULL,
  Xy = NULL,
  nco_offset = rep(0, length(Y)),
  variance = TRUE
)
}
\arguments{
\item{Y}{an n-vector of categorical outcomes with nY distinct values}

\item{offset}{an n-vector of offset for Y}

\item{A}{an n-vector of primary exposure}

\item{X}{an n * nX matrix of adjusted covariates , can be empty}

\item{W}{an n-vector of categorical NCO with nW distinct values}

\item{Z}{an n * nZ matrix of negative control exposure (NCE). Alternatively,
the NCE can be included in the design matrix of the first-stage model by specifying
Xw}

\item{Xw}{a matrix with n rows or a list of length nW, each element is a design matrix of
the corresponding category of W. Default is a matrix with columns of A, Z, X, and Y}

\item{Xw_Y0}{a matrix with n rows or a list of length nW, each element is a design matrix of
the corresponding category of W, with the value of Y fixed at zero. Default is a matrix with columns of A, Z, X,
and Y replaced by zero.}

\item{Xy}{a matrix with n rows or a list of length nY, each element is a design matrix of
the corresponding category of Y besides A, W and linear predictors of W. Default to be X}

\item{nco_offset}{An n-vector of offset for W.}

\item{variance}{whether to return the variance and components for the calculation}
}
\value{
A list with three elements: ESTIMATE includes the parameter estimates from the second-stage
model; SE produces their standard error, and PARAM includes all parameter estimates
}
\description{
This function computes the adjusted odds ratio using multiple logistic regression model,
using a pair of negative control variable
}
\examples{
set.seed(2002)
N <- 2000
expit  <-  function(x) exp(x)/(1 + exp(x))
U <- runif(N, -1, 1); X <- runif(N, -1, 1)
A <- rbinom(N, 1, expit(-3 + 5 * U + 1 * X))
Y <- rep(NA, N)
for (i in 1:N) {
  Yi <- rmultinom(1, 1, c(1,
                          exp(0.2 + 1.5 * U[i] + 0.2 * X[i] + 0.2 * A[i]),
                          exp(0.1 + 3 * U[i] + 0.1 * X[i] + 0.4 * A[i])))
  Y[i] <- which(Yi == 1) - 1
}

W <- rep(NA, N)
for (i in 1:N) {
  W[i] <- rbinom(1, 1, expit(0.1 + 1 * U[i] + 0.1 * X[i]))
}
Z <- rnorm(N, 2 * U + 0.5 * X)

pci_result <- pci.logitreg(Y = Y, A = A, X = X,
       W = W, Z = Z)
pci_result$summary_first_stage
pci_result$summary_second_stage
}
