% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcinb.R
\name{pcinb2s}
\alias{pcinb2s}
\title{Proximal causal inference with count outcomes using two-stage-least-square}
\usage{
pcinb2s(
  Y,
  offset = rep(0, length(Y)),
  A,
  X = NULL,
  W,
  Z = NULL,
  Xw = NULL,
  Xy = NULL,
  nboot = 2000,
  nco_type = NULL,
  nco_args = NULL,
  nb_init = NA,
  se_method = "analytic"
)
}
\arguments{
\item{Y}{an n-vector of count outcomes}

\item{offset}{an n-vector of offset for Y}

\item{A}{an n-vector of primary exposure}

\item{X}{an n * nX matrix of adjusted covariates , can be empty}

\item{W}{an n * nW matrix of negative control outcome (NCO)}

\item{Z}{an n * nZ matrix of negative control exposure (NCE). Alternatively,
the NCE can be included in the design matrix of the first-stage model by specifying
Xw}

\item{Xw}{a matrix with n rows or a list of length nW, each element is a design matrix of
the corresponding NCO. Default is a matrix with columns of A, Z, X, and interactions between
A, Z and A, X}

\item{Xy}{a matrix with n rows, covariates to be adjusted for in the second-stage
model besides A and predictors for W. This may include interaction between A and columns of X.
Default to be X}

\item{nco_type}{An nW-vector for the types of models used in the first-stage, including
"linear" (linear model), "loglin" (log-linear model), "poisson" (Poisson regression model),
"negbin" (negative-binomial regression model), and "ah" (Lin \& Ying's additive hazards model)}

\item{nco_args}{A list of length nW containing additional arguments for the first-stage models.
Each element in the list is a sublist and should contain a vector named "offset" (default 0).
If nco_type == "ah", the sublist needs to include another vector named "event" as the event indicator
(default 1). If nco_type == "negbin", the sublist needs to include a vector "init" as the initial values,
which can be NA.}

\item{se_method}{Method to compute the standard error, can be one of "all", "analytic", "exponential multiplier",
"gaussian multiplier" or "none".}
}
\value{
A list with three elements: ESTIMATE includes the parameter estimates from the second-stage
model; SE produces their standard error, and PARAM includes all parameter estimates
}
\description{
'pcinb2s()' computes the adjusted rate ratio using negative-binomial regression model,
using a pair of negative control variable
}
\examples{
N <- 2000
U <- runif(N); X <- runif(N)
A <- rbinom(N, 1, expit(-3 + 5 * U + 1 * X))
Y <- rpois(N, exp(0.2 + 1.5 * U + 0.2 * X + 0.2 * A))
Z <- rnorm(N, 2 * U + 0.5 * X)
W2 <- rexp(N, 0.1 + 1 * U + 0.1 * X)
D2 <- as.numeric(W2 < 5)
W2[D2 == 0] <- 5
W <- cbind(rnbinom(N, size = 25,
                   mu = exp(2.5 * U + 0.2 * X)),
           W2)
pcinb2s(Y = Y, A = A, X = X,
       W = W, Z = Z, nboot = 100,
       nco_type = c("negbin", "ah"),
       nco_args = list(list(offset = rep(0, N)),
                       list(offset = rep(0, N),
                            event = D2)))
}
