% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglin.R
\name{poisson_fit}
\alias{poisson_fit}
\title{Poisson regression model}
\usage{
poisson_fit(
  y,
  x,
  offset = rep(0, length(y)),
  variance = TRUE,
  interval = c(-100, 100)
)
}
\arguments{
\item{y}{n-vector of outcome}

\item{x}{design matrix}

\item{offset}{n-vector of offset}

\item{variance}{whether to return the variance and components for the calculation}
}
\description{
Fit Poisson regression model Y | X ~ Poisson(exp(offset + beta * x))
}
