% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglin.R
\name{negbin_fit}
\alias{negbin_fit}
\title{Negative-binomial regression model}
\usage{
negbin_fit(
  y,
  x,
  offset = rep(0, length(y)),
  variance = TRUE,
  interval = c(-100, 100),
  init = NA
)
}
\arguments{
\item{y}{n-vector of outcome}

\item{x}{design matrix}

\item{offset}{n-vector of offset}

\item{variance}{whether to return the variance and components for the calculation}

\item{init}{a vector of length ncol(x)+1, initial values for the negative-binomial regression.
Default is crude estimates of the result.}
}
\description{
Fit negative-binomial regression model Y | X ~ NegativeBinomial(theta, exp(offset + beta * x)).
This is a utility function. For formal negative-binomial regression models,
MASS::glm.nb is recommended
}
\examples{
x <- cbind(1, rep(1:3, each = 20))
tt <- rep(1:2, 30)
y <- rnbinom(60, size = 2, mu = tt * exp(x[, 1] + 0.5 * x[, 2]))

negbin_result <- negbin_fit(y, x, offset = log(tt))
negbin_result$summary
}
